catalog/recurring updated

+++ upload/admin/controller/payment/firstdata_remote.php
@@ -408,9 +408,9 @@ class ControllerPaymentFirstdataRemote extends Controller {
 
 				$json['data'] = array();
 				$json['data']['created'] = date("Y-m-d H:i:s");
-				$json['data']['amount'] = (double)$firstdata_order['total'];
+				$json['data']['amount'] = (float)$firstdata_order['total'];
 				$json['data']['capture_status'] = $capture_status;
-				$json['data']['total'] = (double)$total_captured;
+				$json['data']['total'] = (float)$total_captured;
 				$json['data']['total_formatted'] = $this->currency->format($total_captured, $firstdata_order['currency_code'], 1, true);
 				$json['error'] = false;
 			} else {
@@ -467,8 +467,8 @@ class ControllerPaymentFirstdataRemote extends Controller {
 				$json['data'] = array();
 				$json['data']['created'] = date("Y-m-d H:i:s");
 				$json['data']['amount'] = $firstdata_order['total'] * -1;
-				$json['data']['total_captured'] = (double)$total_captured;
-				$json['data']['total_refunded'] = (double)$total_refunded;
+				$json['data']['total_captured'] = (float)$total_captured;
+				$json['data']['total_refunded'] = (float)$total_refunded;
 				$json['data']['refund_status'] = $refund_status;
 				$json['error'] = false;
 			} else {
