+++ upload/admin/controller/payment/realex.php
@@ -429,7 +429,7 @@ class ControllerPaymentRealex extends Controller {
 				$json['data']['created'] = date("Y-m-d H:i:s");
 				$json['data']['amount'] = $this->request->post['amount'];
 				$json['data']['capture_status'] = $capture_status;
-				$json['data']['total'] = (double)$total_captured;
+				$json['data']['total'] = (float)$total_captured;
 				$json['error'] = false;
 			} else {
 				$json['error'] = true;
@@ -484,8 +484,8 @@ class ControllerPaymentRealex extends Controller {
 				$json['data'] = array();
 				$json['data']['created'] = date("Y-m-d H:i:s");
 				$json['data']['amount'] = $this->request->post['amount']*-1;
-				$json['data']['total_captured'] = (double)$total_captured;
-				$json['data']['total_rebated'] = (double)$total_rebated;
+				$json['data']['total_captured'] = (float)$total_captured;
+				$json['data']['total_rebated'] = (float)$total_rebated;
 				$json['data']['rebate_status'] = $rebate_status;
 				$json['error'] = false;
 			} else {
